+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ### Improvements
 
+- feat(nodegroup): add opt to attach extra security groups
+  [#390](https://github.com/pulumi/pulumi-eks/pull/290)
 - feat(cluster): add encryptionConfigKeyArn opt to encrypt k8s Secrets
   [#389](https://github.com/pulumi/pulumi-eks/pull/389)